pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "job-tracker-api"
version = "0.1.0"
description = "Job Tracker API"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111",
  "uvicorn[standard]>=0.30",
  "pydantic[email]>=2.7",
  "pydantic-settings>=2.3",
  "sqlalchemy>=2.0",
  "alembic>=1.13",
  "psycopg[binary]>=3.1",
  "python-jose[cryptography]>=3.3",
  "passlib[argon2,bcrypt]>=1.7",
  "python-multipart>=0.0.9",
  "httpx>=0.27",
  "bcrypt==4.0.1",
  "argon2-cffi>=23.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.5",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = [
  "E501",
  "B008",  # FastAPI Depends() en firmas es OK
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
filterwarnings = [
    "ignore:'crypt' is deprecated:DeprecationWarning",
    "ignore:Accessing argon2.__version__ is deprecated:DeprecationWarning",
]
